### 2024-10-21 #7387 @editorialbot
[PRE REVIEW]: Photon Weave
**Submitting author:** @SimonSekavcnik (<a href="http://orcid.org/0000-0002-1370-9751">Simon Sekavcnik</a>)
**Repository:** https://github.com/tqsd/photon_weave
**Branch with paper.md** (empty if default branch): master
**Version:** 0.1.4
