### 2024-11-12 #7387 @a-eghrari
Hi, I am available to review this paper this paper.Best,AliOn Nov 11, 2024, at 7:40 PM, Daniel S. Katz ***@***.***> wrote:﻿
👋 @a-eghrari & @taladjidi & @PhilipVinc, would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
