### 2024-11-22 4d2d0ca
Lower number of shots for MZ example to be kinder to potato laptops
