### 2024-11-22 dc8d5a5
Avoid pure python loops and dictionnaries when everything uses JAX / Numpy arrays
