+++ examples/mach_zehnder_interferometer.py
@@ -10,7 +10,19 @@ from photon_weave.state.composite_envelope import CompositeEnvelope
 from photon_weave.state.envelope import Envelope
 
 
-def mach_zender_single_shot(phase_shift: float):
+def mach_zender_single_shot(phase_shift: float) -> list[int]:
+    """Return photon count in each port of a Mach-Zehnder Interferometer.
+
+    Parameters
+    ----------
+    phase_shift : float
+        Phase shift between the two arms of the interferometer.
+
+    Returns
+    -------
+    list[int]
+        Photon count in each port of the interferometer.
+    """
     # Create one envelope
     env1 = Envelope()
     # Create one photon
@@ -36,23 +48,21 @@ def mach_zender_single_shot(phase_shift: float):
 
 if __name__ == "__main__":
     from tqdm import tqdm
+    import numpy as np
     num_shots = 100
-    angles = jnp.linspace(0, 2 * jnp.pi, 25)
-    results = {float(angle): [] for angle in angles}
+    angles = jnp.linspace(0, 2 * jnp.pi, 10)
+    # (num_angles, num_shots , num_ports)
+    results = np.zeros((angles.shape[0], num_shots, 2))
     pbar = tqdm(total=len(angles) * num_shots, desc="Simulating Mach-Zehnder Interferometer")
-    for angle in angles:
-        for _ in range(num_shots):
+    for i, angle in enumerate(angles):
+        for j in range(num_shots):
             shot_result = mach_zender_single_shot(angle)
-            results[float(angle)].append(shot_result)
+            # results[float(angle)].append(shot_result)
+            results[i, j, :] = shot_result
             pbar.update(1)
     pbar.close()
-    measurements_1 = []
-    measurements_2 = []
-    for angle, shots in results.items():
-        counts_1 = sum(shot[0] for shot in shots) / num_shots
-        counts_2 = sum(shot[1] for shot in shots) / num_shots
-        measurements_1.append(counts_1)
-        measurements_2.append(counts_2)
+    measurements_1 = results.mean(axis=1)[:, 0]
+    measurements_2 = results.mean(axis=1)[:, 1]
 
     # Plot results
     plt.figure(figsize=(10, 6))
