### 2024-11-22 #7468 @taladjidi
Hi @danielskatz , I only just started my review but here are some remarks.

**Paper**:
- l.12 "they often require meticulous operations organization for complex simulations that might require professional experience in software skills." I'm not sure this is a fair statement for software designed for physicits. Furthermore, while I'm familiar with QuTip, I've never used it for involved simulation, however judging from their examples it exposes a similar API to your package. See [here](https://nbviewer.org/urls/qutip.org/qutip-tutorials/tutorials-v5/time-evolution/017_smesolve-jc-photocurrent.ipynb) for an example of a simulation of a photodiode current with stochastic decay.
- l.15 to l.21 When comparing to other frameworks, the claims should be more substantiated with actual comparisons. Especially concerning speed / performance. II would advise a benchmarking example with the same simple case treated with different frameworks. The MZ interferometer should do the trick.
- How does your package compare to [SOQCS](https://github.com/SOQCSAdmin/SOQCS) ?
- l.29 "fostering greater collaboration ..." it seems that there are a lot of people using circuit representations of quantum optical setups as a way to simplify them. Furthermore this representation is a very convenient tool for computations with entanglement propagation / computations with photon cluster states etc ... Thus I would refrain from implying that going away from this representation fosters collaboration. On the contrary, a conversion utility from circuit representation to your framework would indeed add more flexibilty and foster collaboration through interoperability.
- l.38: I understand from this that your package cannot simulate continuous wave setups (for instance typical quantum optics experiments like squeezing). Moreover this only simulates linear operations ?
- l.47 : "tensorized" I had never encountered this term. While I understand (correct me if I'm wrong) that it designates the conversion of one state representation to a tensor, it would be better to explicit this a bit more.
- l.53: "temporal profile" I understand that this is only relevant for pulses, maybe it would be good to mention it earlier that the package simulates pulses.
- l.67: I would avoid including source code in the paper unless it is to highlight the ease of use of your package compared to the other packages. This could be a good figure if for instance you compare your operation definition to another framework ?
- l.68: Resource utilization claims should be substantiated quantitatively.
- l.74 "quantum channels defined by Kraus operators" it should be explicited earlier what your concept of a "channel" encapsulates. This is explained well in your docs.
- l.76 The description of the measurement is a bit short while it is described extensively in your docs. I would maybe include a figure presenting symbolically the different measurement types.
- l.84 "interoperable quantum devices maybe provide an example of this interoperability ?

**Code**:
- I've drafted a PR with preliminary comments and edits [here](https://github.com/tqsd/photon_weave/pull/8)
- Documentation is a bit light in figures and examples. There are some typos in the docs ("Reproducability" for instance).
- There is no contribution guide
- There are not a lot of examples, especially compared to the packages that you compare with PhotonWeave.
- There is not indication of how to run the tests.

Overall, I expect a clearer / more visual demonstration of what your software is capable of. Figures will help a lot in conveying your message in a convincing manner.

Don't hesitate to ping me if you need help, I'll try to do what I can to be the most helpful possible. Having appreciated how nice and wholesome the review process is at JOSS, I would like to do everything I can to keep it that way 👼
