### 2024-11-22 b4fe228
Comment about performance
