### 2024-11-22 #7468 @SimonSekavcnik
Thanks a lot for your reveiw so far @taladjidi. I will process your pull requests and attend the rest of the topics you have addressed.

@danielskatz how should I proceed, should I work on the updates in the separate branch or can I update the master branch during the review process. For example, can I update the paper before I get the second review?
