### 2024-11-24 #7468 @danielskatz
👋 @a-eghrari - Are you able to start your review? Please see https://github.com/openjournals/joss-reviews/issues/7468#issuecomment-2470662829 for how generate your checklist, after which you can start checking items off as you review the submission.
