### 2024-11-27 #7468 @SimonSekavcnik
@taladjidi, We have given some comments to the pull request. After they are cleared we would like to merge the request into the master branch. The comments were insightful and we will address all of the proposed fixes in the commits following the pull request merge.

According to your review, we would like to add more examples of usage. We have an example of the Jaynes-Cummings model, like the one you linked in the initial comments. Furthermore we would like to create another example of the photonic processor. If you would like to, you can propose any other example.

In regards to the benchmarking, my coauthor is working on comparisons with the qutip and strawberry fields. Do you think that will suffice, or should we include comparisons to a larger set of frameworks.

I will update the paper according to your comments and submit it as a pull request to the master branch.
