### 2024-11-27 f3219c8
Adjusted dimension estimation in the CompositeOperation type
