### 2024-11-27 da97e10
Adjusted the test to reflect the previous change
