### 2024-11-27 3d7ef14
Added Jaynes-Cummings Model
