+++ photon_weave/state/fock.py
@@ -460,7 +460,7 @@ class Fock(BaseState):
             if not jnp.any(jnp.abs(self.state) > 0):
                 raise ValueError(
                     "The state is entirely composed of zeros, is |0⟩"
-                    " attempted to be anniilated?"
+                    " attempted to be annihilated?"
                 )
             if operation.renormalize:
                 self.state = self.state / jnp.linalg.norm(self.state)
