+++ photon_weave/state/fock.py
@@ -20,6 +20,8 @@ from photon_weave.state.composite_envelope import CompositeEnvelope
 
 from .base_state import BaseState
 from .expansion_levels import ExpansionLevel
+from .utils.measurements import measure_vector, measure_matrix
+from .utils.routing import route_operation
 
 if TYPE_CHECKING:
     from .envelope import Envelope
@@ -60,9 +62,8 @@ class Fock(BaseState):
     __slots__ = (
         "uid",
         "index",
-        "label",
+        "state",
         "dimensions",
-        "state_vector",
         "density_matrix",
         "envelope",
         "expansions",
@@ -106,6 +107,7 @@ class Fock(BaseState):
                     return True
         return False
 
+    @route_operation()
     def expand(self) -> None:
         """
         Expands the representation. If the state is stored in
@@ -113,12 +115,12 @@ class Fock(BaseState):
         state is in state_vector, then the state is expanded
         to the state_matrix
         """
-        if isinstance(self.index, int):
-            assert isinstance(self.envelope, Envelope)
-            self.envelope.expand()
-        if isinstance(self.index, tuple) or isinstance(self.index, list):
-            assert isinstance(self.composite_envelope, CompositeEnvelope)
-            self.composite_envelope.expand(self)
+        #if isinstance(self.index, int):
+        #    assert isinstance(self.envelope, Envelope)
+        #    self.envelope.expand()
+        #if isinstance(self.index, tuple) or isinstance(self.index, list):
+        #    assert isinstance(self.composite_envelope, CompositeEnvelope)
+        #    self.composite_envelope.expand(self)
 
         if self.dimensions < 0:
             assert self.dimensions is not None, "self.dimensions shoul not be None"
@@ -245,6 +247,7 @@ class Fock(BaseState):
         else:
             self.index = minor
 
+    @route_operation()
     def measure(
         self, separate_measurement: bool = False, destructive: bool = True
     ) -> Dict[BaseState, int]:
@@ -266,42 +269,20 @@ class Fock(BaseState):
         Dict[BaseState, int]
             Dictionary of outcomes
         """
-        if isinstance(self.index, int):
-            assert isinstance(self.envelope, Envelope)
-            return self.envelope.measure(
-                self, separate_measurement=separate_measurement, destructive=destructive
-            )
-        if isinstance(self.index, tuple) or isinstance(self.index, list):
-            assert isinstance(self.composite_envelope, CompositeEnvelope)
-            return self.composite_envelope.measure(self)
-
-        if self.index is not None:
-            assert self.envelope is not None, "Envelope should not be None"
-            return self.envelope.measure()
-        C = Config()
         match self.expansion_level:
             case ExpansionLevel.Label:
                 assert isinstance(self.state, int)
-                result = self.state
+                outcomes = {self:self.state}
             case ExpansionLevel.Vector:
-                assert isinstance(self.state, jnp.ndarray)
-                assert self.state.shape == (self.dimensions, 1)
-                probs = jnp.abs(self.state.flatten()) ** 2
-                probs = probs.ravel()
-                assert jnp.isclose(sum(probs), 1)
-                key = C.random_key
-                result = int(jax.random.choice(key, a=jnp.arange(len(probs)), p=probs))
+                outcomes, post_measurement_state = measure_vector(
+                    [self], [self], self.state)
             case ExpansionLevel.Matrix:
-                assert isinstance(self.state, jnp.ndarray)
-                assert self.state.shape == (self.dimensions, self.dimensions)
-                probs = jnp.diag(self.state).real
-                probs = probs / jnp.sum(probs)
-                key = C.random_key
-                result = int(jax.random.choice(key, a=jnp.arange(len(probs)), p=probs))
-        self.state = result
[…]
