### 2024-11-29 a108ae8
Abstracted the operation application and cleaned Fock
