### 2024-11-29 8dec41f
Cleaned up Polarization
