+++ photon_weave/state/polarization.py
@@ -17,6 +17,9 @@ from photon_weave.photon_weave import Config
 
 from .base_state import BaseState
 from .expansion_levels import ExpansionLevel
+from .utils.measurements import measure_matrix, measure_vector
+from .utils.operations import apply_operation_matrix, apply_operation_vector
+from .utils.routing import route_operation
 
 if TYPE_CHECKING:
     from photon_weave.operation import Operation
@@ -77,8 +80,6 @@ class Polarization(BaseState):
         polarization: PolarizationLabel = PolarizationLabel.H,
         envelope: Union["Envelope", None] = None,
     ):
-        # from photon_weave.state.composite_envelope import CompositeEnvelope
-        # from photon_weave.state.envelope import Envelope
 
         self.uid: uuid.UUID = uuid.uuid4()
         logger.info("Creating polarization with uid %s", self.uid)
@@ -265,6 +266,7 @@ class Polarization(BaseState):
         self.index = None
         self.expansion_level = None
 
+    @route_operation()
     def measure(
         self, separate_measurement: bool = False, destructive: bool = True
     ) -> Dict[BaseState, int]:
@@ -277,58 +279,44 @@ class Polarization(BaseState):
         separate_measurement: bool
 
         Returns
-        ----
+        -------
         Union[int,None]
             Measurement Outcome
+
+        Notes
+        -----
+        Method is decorated with route_operation. If the state is
+        contained in the product state, the corresponding operation
+        will be executed in the state container, which contains this
+        stat.
         """
         from photon_weave.state.composite_envelope import CompositeEnvelope
 
-        # If the state is in the envelope, measure there
-        if isinstance(self.index, int):
-            assert isinstance(self.envelope, Envelope)
-            return self.envelope.measure(
-                self, separate_measurement=separate_measurement, destructive=destructive
-            )
-
-        # If the state is in the composite envelope, measure there
-        if isinstance(self.index, tuple) or isinstance(self.index, list):
-            assert isinstance(self.composite_envelope, CompositeEnvelope)
-            return self.composite_envelope.measure(self)
-
-        results: Dict[BaseState, int] = {}
-        C = Config()
         if self.expansion_level == ExpansionLevel.Label:
             self.expand()
-        if self.expansion_level == ExpansionLevel.Vector:
-            assert isinstance(self.state, jnp.ndarray)
-            assert self.state.shape == (self.dimensions, 1)
-            prob_0 = jnp.abs(self.state[0]) ** 2
-            prob_1 = jnp.abs(self.state[1]) ** 2
-            assert jnp.isclose(prob_0 + prob_1, 1.0)
-            probs = jnp.array([prob_0[0], prob_1[0]])
-            key = C.random_key
-            outcome = jax.random.choice(key, a=jnp.array([0, 1]), p=probs.ravel())
-            results[self] = int(outcome)
-        elif self.expansion_level == ExpansionLevel.Matrix:
-            assert isinstance(self.state, jnp.ndarray)
-            assert self.state.shape == (self.dimensions, self.dimensions)
-            probabilities = jnp.diag(self.state).real
-            probabilities = probabilities / jnp.sum(probabilities)
-            # Generate a random key
-            key = C.random_key
-            outcome = jax.random.choice(
-                key, a=jnp.arange(self.state.shape[0]), p=probabilities
-            )
-            results[self] = int(outcome)
-        if results[self] == 0:
+
+        match self.expansion_level:
+            case ExpansionLevel.Vector:
+                outcomes, post_measurement_state = measure_vector(
+                    [self],[self], self.state
+                    )
+            case ExpansionLevel.Matrix:
+                outcomes, post_measurement_state = measure_matrix(
+                    [self],[self], self.state
+                    )
+        # Reconstruct the state post measurement
+        if outcomes[self] == 0:
[…]
