+++ photon_weave/_math/ops.py
@@ -372,7 +372,8 @@ def kraus_identity_check(
     operators: List[Union[np.ndarray, jnp.ndarray]], tol: float = 1e-6
 ) -> bool:
     """
-    Check if Kraus operators sum to the identity matrix.
+    Check if Kraus operator sum is less or equal to identity,
+    thus representing a valid Kraus Channel
 
     Parameters
     ----------
@@ -388,8 +389,11 @@ def kraus_identity_check(
     """
     dim = operators[0].shape[0]
     identity_matrix = jnp.eye(dim)
-    sum_kraus = sum(jnp.matmul(jnp.conjugate(K.T), K) for K in operators)
-    return jnp.allclose(sum_kraus, identity_matrix, atol=tol).item()
+    kraus_sum = sum(jnp.matmul(jnp.conjugate(K.T), K) for K in operators)
+    identity = jnp.eye(kraus_sum.shape[0])
+    is_valid = jnp.all(jnp.real(jnp.linalg.eigvals(identity-kraus_sum)) >= 0)
+    #return jnp.allclose(sum_kraus, identity_matrix, atol=tol).item()
+    return is_valid
 
 
 @jit
