+++ photon_weave/state/custom_state.py
@@ -18,6 +18,9 @@ from photon_weave.photon_weave import Config
 from photon_weave.state.base_state import BaseState
 from photon_weave.state.composite_envelope import CompositeEnvelope
 from photon_weave.state.expansion_levels import ExpansionLevel
+from .utils.measurements import measure_vector, measure_matrix
+from .utils.operations import apply_operation_vector, apply_operation_matrix
+from .utils.routing import route_operation
 
 
 class CustomState(BaseState):
@@ -61,9 +64,18 @@ class CustomState(BaseState):
         assert not isinstance(index, int)
         self._index = index
 
+
+    @route_operation()
     def expand(self) -> None:
         """
         Expands the state from label to vector and from vector to matrix
+        
+        Notes
+        -----
+        Method is decorated with route_operation. If the state is
+        contained in the product state, the corresponding operation
+        will be executed in the state container, which contains this
+        stat.
         """
         if isinstance(self.index, tuple):
             assert isinstance(self.composite_envelope, CompositeEnvelope)
@@ -161,6 +173,7 @@ class CustomState(BaseState):
                 "Either set both parameters (minor, major) or none of them"
             )
 
+    @route_operation()
     def measure(
         self, separate_measurement: bool = False, destructive: bool = True
     ) -> Dict[BaseState, int]:
@@ -180,48 +193,31 @@ class CustomState(BaseState):
         -------
         Dict[BaseState, int]
             Dictionary of outcomes, in this case only one outcome
+
+        Notes
+        -----
+        Method is decorated with route_operation. If the state is
+        contained in the product state, the corresponding operation
+        will be executed in the state container, which contains this
+        stat.
         """
-        if self.index is not None:
-            assert self.composite_envelope is not None
-            return self.composite_envelope.measure(self)
-        elif self.index is None:
-            if self.expansion_level == ExpansionLevel.Label:
+        match self.expansion_level:
+            case ExpansionLevel.Label:
                 assert isinstance(self.state, int)
                 return {self: self.state}
-            elif self.expansion_level == ExpansionLevel.Vector:
-                assert isinstance(self.state, jnp.ndarray)
-                assert self.state.shape == (self.dimensions, 1)
-                C = Config()
-                probabilities = jnp.abs(self.state.flatten()) ** 2
-                probabilities = probabilities.ravel()
-                assert jnp.isclose(sum(probabilities), 1)
-                key = C.random_key
-                out = int(
-                    jax.random.choice(
-                        key, a=jnp.array(len(probabilities)), p=probabilities
-                    )
+            case ExpansionLevel.Vector:
+                outcomes, post_measurement_state = measure_vector(
+                    [self], [self], self.state
                 )
-                self.state = out
-                self.expansion_level = ExpansionLevel.Label
-                return {self: out}
-            elif self.expansion_level == ExpansionLevel.Matrix:
-                assert isinstance(self.state, jnp.ndarray)
-                assert self.state.shape == (self.dimensions, self.dimensions)
-                C = Config()
-                probabilities = jnp.diag(self.state).real
-                probabilities = probabilities / jnp.sum(probabilities)
-                key = C.random_key
-                out = int(
-                    jax.random.choice(
-                        key, a=jnp.array(len(probabilities)), p=probabilities
-                    )
+            case ExpansionLevel.Matrix:
+                outcomes, post_measurement_state = measure_matrix(
+                    [self], [self], self.state
                 )
-                self.state = out
-                self.expansion_level = ExpansionLevel.Label
-                return {self: out}
-        raise ValueError(
-            "Something went wrong, this exception should not be raised"
-        )  # pragma: no cover
[…]
