### 2024-11-29 #7468 @taladjidi
Hi @SimonSekavcnik , I'll look through the PR again this weekend or beginning of next week, I was quite overbooked this week. I think comparing with qutip and strawberry fields should suffice. Did you have a look at SOQCS ?
