### 2024-11-29 #7468 @SimonSekavcnik
Hi @taladjidi, no worries, I am addressing your comments in my fork already. I think the update will lead to a lot more readable code. Your comments are really appreciated.

I took a look yes. I wasn't aware of it before. I will definitely use it to help test some of our models. I didn't have a lot of time to play around with it, but I wasn't able to find how one could define custom devices. Further more, in our models we explicitly include polarization, which is not part of the SOQCS as far as I was able to see it. Another (important) difference is the fact that we include functionality to model interaction between the light modes and arbitrary quantum system (as long as user can define the operators). Here I am not sure if the comparison is strictly necessary in the paper, since we are severely limited by the word count.

Maybe I can add another note here touching on the interoperability. If you want to create a component which will operate on one system, you as the user, don't need to understand how the product state is constructed. As you can apply the operations to the states which pertain to your component, the operators will be applied for you to the correct subspace of the product space. Maybe this feature was not communicated in the best way, but I will try to illuminate what I mean by so examples.
