### 2024-11-29 eca5b16
Implemented expand, contract and used it in base states
