+++ photon_weave/state/polarization.py
@@ -20,6 +20,7 @@ from .expansion_levels import ExpansionLevel
 from .utils.measurements import measure_matrix, measure_vector
 from .utils.operations import apply_operation_matrix, apply_operation_vector
 from .utils.routing import route_operation
+from .utils.state_transform import state_expand, state_contract
 
 if TYPE_CHECKING:
     from photon_weave.operation import Operation
@@ -39,6 +40,8 @@ class PolarizationLabel(Enum):
     V = "V"
     R = "R"
     L = "L"
+    A = "A"
+    D = "D"
 
 
 class Polarization(BaseState):
@@ -109,26 +112,21 @@ class Polarization(BaseState):
             "Dimensions can not be set for Polarization type, 2 by default"
         )
 
+    @route_operation()
     def expand(self) -> None:
         """
         Expands the representation
         If current representation is label, then it gets
         expanded to state_vector and if it is state_vector
         then it gets expanded to density matrix
+
+        Notes
+        -----
+        Method is decorated with route_operation. If the state is
+        contained in the product state, the corresponding operation
+        will be executed in the state container, which contains this
+        state.
         """
-        from photon_weave.state.composite_envelope import CompositeEnvelope
-
-        # If the state is in composite envelope expand the product space there
-        if isinstance(self.index, tuple) or isinstance(self.index, list):
-            assert isinstance(self.composite_envelope, CompositeEnvelope)
-            self.composite_envelope.expand(self)
-            return
-        # If the state is in envelope expand the product space there
-        elif isinstance(self.index, int):
-            assert isinstance(self.envelope, Envelope)
-            self.envelope.expand()
-            return
-
         if self.expansion_level == ExpansionLevel.Label:
             assert isinstance(self.state, PolarizationLabel)
             vector: List[Union[jnp.ndarray, float, complex]]
@@ -143,14 +141,20 @@ class Polarization(BaseState):
                 case PolarizationLabel.L:
                     # Left circular polarization = (1/sqrt(2)) * (|H⟩ - i|V⟩)
                     vector = [1 / jnp.sqrt(2), -1j / jnp.sqrt(2)]
+                case PolarizationLabel.A:
+                    vector = [1 / jnp.sqrt(2), -1 / jnp.sqrt(2)]
+                case PolarizationLabel.D:
+                    vector = [1 / jnp.sqrt(2), 1 / jnp.sqrt(2)]
             self.state = jnp.array(vector)[:, jnp.newaxis]
             self.expansion_level = ExpansionLevel.Vector
-        elif self.expansion_level == ExpansionLevel.Vector:
-            assert isinstance(self.state, jnp.ndarray)
-            assert self.state.shape == (self.dimensions, 1)
-            self.state = jnp.dot(self.state, jnp.conj(self.state.T))
-            self.expansion_level = ExpansionLevel.Matrix
+        else:
+            self.state, self.expansion_level = state_expand(
+                self.state,
+                self.expansion_level,
+                self.dimensions
+                )
 
+    @route_operation()
     def contract(
         self, final: ExpansionLevel = ExpansionLevel.Label, tol: float = 1e-6
     ) -> None:
@@ -163,45 +167,28 @@ class Polarization(BaseState):
             Expected expansion level after contraction
         tol: float
             Tolerance when comparing matrices
+
+        Notes
+        -----
+        Method is decorated with route_operation. If the state is
+        contained in the product state, the corresponding operation
+        will be executed in the state container, which contains this
+        state.
         """
-        from photon_weave.state.envelope import Envelope
+        #from photon_weave.state.envelope import Envelope
 
         # If state was measured, then do nothing
         if self.measured:
-            return
[…]
