### 2024-12-02 374b13e
Cleaned the representation functionality
