### 2024-12-02 88b554e
Cleaned up measurement logic in envelope
