+++ photon_weave/state/envelope.py
@@ -30,6 +30,8 @@ from photon_weave.state.polarization import Polarization
 
 from .utils.state_transform import state_contract, state_expand
 from .utils.representation import representation_vector, representation_matrix
+from .utils.measurements import measure_vector, measure_matrix
+from .utils.trace_out import trace_out_vector, trace_out_matrix
 
 if TYPE_CHECKING:
     from photon_weave.operation import Operation
@@ -274,251 +276,58 @@ class Envelope:
         # Check if given states are part of this envelope
         for s in states:
             assert s in [self.fock, self.polarization]
+        if len(states) == 0:
+            states = (self.fock, self.polarization)
 
         outcomes = {}
-        reshape_shape = []
         if self.state is None:
             for s in [self.polarization, self.fock]:
                 out = s.measure()
                 for k, v in out.items():
                     outcomes[k] = v
         else:
-            assert isinstance(self.fock.index, int)
-            assert isinstance(self.polarization.index, int)
+            self.reorder(self.fock, self.polarization)
+            if not separate_measurement and len(states) == 1:
+                states = [self.fock, self.polarization]
+            if len(states) == 2:
+                separate_measurement = True
 
-            reshape_shape = [-1, -1]
-            reshape_shape[self.fock.index] = self.fock.dimensions
-            reshape_shape[self.polarization.index] = self.polarization.dimensions
+            assert all(isinstance(s.index, int) for s in states)
 
-            C = Config()
+            match self.expansion_level:
+                case ExpansionLevel.Vector:
+                    outcomes, self.state = measure_vector(
+                        [self.fock, self.polarization],
+                        states,
+                        self.state)
+                case ExpansionLevel.Matrix:
+                    outcomes, self.state = measure_matrix(
+                        [self.fock, self.polarization],
+                        states,
+                        self.state)
 
-            if self.expansion_level == ExpansionLevel.Vector:
-                assert isinstance(self.state, jnp.ndarray)
-                assert self.state.shape == (self.dimensions, 1)
-                reshape_shape.append(1)
-                ps = self.state.reshape(reshape_shape)
-
-                # 1. Measure Fock Part
-                if (
-                    (separate_measurement and self.fock in states)
-                    or len(states) == 0
-                    or len(states) == 2
-                ):
-                    probabilities = (
-                        jnp.abs(jnp.sum(ps, axis=self.polarization.index)).flatten()
-                        ** 2
-                    )
-                    key = C.random_key
-                    choice = int(
-                        jax.random.choice(
-                            key, a=jnp.arange(len(probabilities)), p=probabilities
-                        )
-                    )
-                    outcomes[self.fock] = choice
-
-                    # Construct post measurement state
-                    post_measurement = jnp.take(ps, choice, self.polarization.index)
-                    ps = jnp.take(ps, choice, axis=self.fock.index)
-
-                    einsum = "ij,kj->ikj"
-                    if self.fock.index == 0:
-                        ps = jnp.einsum(einsum, post_measurement, ps)
-                    elif self.fock.index == 1:
-                        ps = jnp.einsum(einsum, ps, post_measurement)
-
-                if (
-                    (separate_measurement and self.polarization in states)
-                    or len(states) == 0
-                    or len(states) == 2
-                ):
-                    probabilities = (
-                        jnp.abs(jnp.sum(ps, axis=self.fock.index)).flatten() ** 2
-                    )
-                    key = C.random_key
-                    choice = int(
-                        jax.random.choice(
-                            key, a=jnp.arange(len(probabilities)), p=probabilities
-                        )
-                    )
-                    outcomes[self.polarization] = choice
-
-                    # Construct post measurement state
-                    post_measurement = jnp.take(ps, choice, self.polarization.index)
[…]
