### 2024-12-02 472dc6c
Cleaned up the operation application in Envelope
