### 2024-12-03 b91d5b5
Envelope, cleaned up the POVM measurements
