### 2024-12-03 695cef1
Removed logging from envelope
