### 2024-12-05 4d9ca7b
Cleaned kraus method in envelope
