+++ photon_weave/state/utils/operations.py
@@ -172,17 +172,17 @@ def apply_kraus_vector(state_objs: List[BaseState], target_states: List[BaseStat
         ))
 
     shape = [s.dimensions for s in state_objs]
-    shape.append(1)
+    #shape.append(1)
     
     assert product_state.shape == (dims,1)
-    assert all(operator.shape == (operator_shape, operator_shape) for
+    expected_operator_dims = int(jnp.prod(operator_shape))
+    assert all(operator.shape == (expected_operator_dims, expected_operator_dims) for
                operator in operators)
 
-    product_state = product_state.reshape(shape)
+    product_state = product_state.reshape((*shape,1))
     #operators = [o.reshape((*operator_shape, *operator_shape)) for o in operators]
 
 
-    einsum_o = ESC.apply_operator_vector(state_objs, target_states)
 
     resulting_state = jnp.zeros_like(product_state)
 
@@ -192,17 +192,17 @@ def apply_kraus_vector(state_objs: List[BaseState], target_states: List[BaseStat
     padding_condition = True
     for state in state_objs:
         if state not in target_states:
-            if states.index(state) < states.index(target_states[0]):
-                pre_pad = jnp.kron(pre_pad, jnp.eye(current_element.dimensions))
-            elif states.index(state) > states.index(target_states[-1]):
-                post_pad = jnp.kron(post_pad, jnp.eye(current_element.dimensions))
+            if state_objs.index(state) < state_objs.index(target_states[0]):
+                pre_pad = jnp.kron(pre_pad, jnp.eye(state.dimensions))
+            elif state_objs.index(state) > state_objs.index(target_states[-1]):
+                post_pad = jnp.kron(post_pad, jnp.eye(state.dimensions))
             
-    
+    einsum_o = ESC.apply_operator_vector(state_objs, state_objs)
 
     for operator in operators:
         operator = jnp.kron(pre_pad, jnp.kron(operator, post_pad))
         # We need to expand the operator to affect the whole space
-        operator = operator.reshape((*operator_shape, *operator_shape))
+        operator = operator.reshape((*shape, *shape))
         resulting_state += jnp.einsum(einsum_o, operator, product_state)
 
     resulting_state = resulting_state.reshape((-1,1))
@@ -252,7 +252,8 @@ def apply_kraus_matrix(state_objs: List[BaseState], target_states: List[BaseStat
     shape.append(1)
     
     assert product_state.shape == (dims,dims)
-    assert all(operator.shape == (operator_shape, operator_shape) for
+    expected_operator_dims = int(jnp.prod(operator_shape))
+    assert all(operator.shape == (expected_operator_dims, expected_operator_dims) for
                operator in operators)
 
     product_state = product_state.reshape(shape)
@@ -267,10 +268,10 @@ def apply_kraus_matrix(state_objs: List[BaseState], target_states: List[BaseStat
     padding_condition = True
     for state in state_objs:
         if state not in target_states:
-            if states.index(state) < states.index(target_states[0]):
-                pre_pad = jnp.kron(pre_pad, jnp.eye(current_element.dimensions))
-            elif states.index(state) > states.index(target_states[-1]):
-                post_pad = jnp.kron(post_pad, jnp.eye(current_element.dimensions))
+            if state_objs.index(state) < state_objs.index(target_states[0]):
+                pre_pad = jnp.kron(pre_pad, jnp.eye(state.dimensions))
+            elif state_objs.index(state) > state_objs.index(target_states[-1]):
+                post_pad = jnp.kron(post_pad, jnp.eye(state.dimensions))
             
     
     for operator in operators:
