### 2024-12-05 7cd6a09
simplified contraction/expansion in composite envelope
