### 2024-12-06 1002b58
Cleaned measurement in the composite_envelope
