+++ photon_weave/state/utils/measurements.py
@@ -80,6 +80,7 @@ def measure_matrix(state_objs: List[BaseState], target_states: List[BaseState],
     ----------
     state_objs: List[BaseState]
         List of all state objects which are in the probuct state
+
     states: List[BaseState]
         List of the states, which should be measured
 
