### 2024-12-06 2e42cef
Cleaned up kraus apply in ce
