### 2024-12-06 71b22a7
Cleaned up apply_operation in CE
