+++ photon_weave/state/composite_envelope.py
@@ -242,8 +242,6 @@ class ProductState:
             states,
             operators,
             self.state)
-        print("STATE AFTER POVM")
-        print(self.state)
 
         other_outcomes = {}
         if destructive:
@@ -469,7 +467,8 @@ class ProductState:
             to = states[0].trace_out()
             assert isinstance(to, jnp.ndarray)
             operation.compute_dimensions(states[0]._num_quanta, to)
-            states[0].resize(operation.dimensions[0])
+            succ = states[0].resize(operation.dimensions[0])
+
         elif isinstance(operation._operation_type, PolarizationOperationType):
             assert isinstance(states[0], Polarization)
             assert len(states) == 1
@@ -534,7 +533,7 @@ class CompositeEnvelopeContainer:
 
     def append_states(self, other: "CompositeEnvelopeContainer") -> None:
         """
-        Appnds the states of two composite envelope containers
+        Appends the states of two composite envelope containers
         Parameters
         ----------
         other: CompositeEnvelopeContainer
