### 2024-12-17 #7468 @SimonSekavcnik
In my pull request I have addressed the issues relating to the code. We are still ironing out the issues with the paper. I was delaying the paper update until the second review.
