### 2024-12-25 #7468 @a-eghrari
# Review of Paper Submitted to JOSS

- There are some typos in both the paper and the documentation.
- I have also read @taladjidi's review and PR. I agree with most of the comments, especially regarding code duplication, imports within functions, and formatting of the source code.

## The Paper:
1. **Summary:**
   It would be better to have a shorter summary and move most of the discussion comparing the library with QuTiP and Qiskit to the *Statement of Need* section. While differences can be briefly mentioned in the summary, the detailed comparisons would be more appropriate in *Statement of Need*. The summary should primarily focus on the problem that motivated the creation of this package, what has been achieved, and what results or improvements this work brings.

2. **Computational Efficiency Claims:**
   To support your claim of superior computational efficiency compared to other packages, it would be beneficial to include benchmarking data or showcase how using this package reduces development time or provides another measurable benefit. If you have concerns about the paper's word or space limitations, you could include these details in the documentation and reference them in the paper.

3. **Line 51 Issue:**
   On line 51, in `$\mathcal{F} \otimes \mathcal{P}$`, neither $\mathcal{F}$ nor $\matchless{P}$ has been defined prior to their usage. It would be better to define them, even if their meanings seem clear from the context.

4. **Code in the Paper:**
   The code snippet included in the paper appears unnecessary. Given the word and page limitations, this space could be better used to discuss other important aspects of the package.

---

## The Documentation:
1. **Community Guidelines:**
   There are no community guidelines included.

2. **Lack of Examples:**
   The documentation lacks sufficient examples, which are crucial for understanding and usability. For instance:
   - Examples could be added to clarify concepts like Fock Space, Polarization Space, and Custom Hilbert Space. For example, showcasing the usage of `expand()` or `contract()` methods would be helpful.
   - Similarly, the concepts of Envelopes and Composite Envelopes would benefit from some good examples.

3. **Issues with "Applying Operations" Section:**
   - The code in this section contains a typo in `CustomStateOperaitonType`, which results in an error.
   - There are broader issues with the accessibility of methods. For example, importing
     ```python
     from photon_weave.operation import (
         Operation, FockOperationType,
         PolarizationOperationType,
         CustomStateOperaitonType,
         CompositeOperationType)
     ```
     does not provide access to the `Fock` method as expected. Additionally, the `__init__.py` file in the `state` folder is empty, as is the one in the `photon_weave` folder. Even adding `Fock` and `Polarization` to the imports does not resolve these errors. Please review this example.

4. **Improved Examples of Operations:**
   Instead of presenting all possible operations in a single block, it would be more effective to demonstrate each operation within the context of a small, possible application. This would showcase both the functionality and practical applications of the package.

5. **Existing Examples:**
   The repository contains two examples, but it would be better if these were integrated into the documentation as tutorials/examples.

6. **Issues with Basic Example in the Usage Guide:**
   The basic example provided in the *Usage Guide* does not work. Additionally, there seems to be no source code related to quantum circuits, leading me to assume that this functionality is still in development (Is it?). Perhaps this incomplete example could be replaced with one of the working examples already available in the repository.

---

## Final Thoughts:
I believe the idea behind this work is solid, and the paper is generally well-written, introducing some basic aspects of the package. However, the claims made in the summary and *Statement of Need* require support.

The documentation, while providing good explanations, lacks examples, which are critical for software documentation.
