### 2024-12-30 87d0161
Removed Typo from the documentation and fixed the POVM measurement description. openjournals/joss-reviews#7468
