### 2024-12-30 #7468 @SimonSekavcnik
Thank you both for your insights. We are working on the performance benchmarks. As I have stated before the code duplication problem should be addressed in my pull request. I will add some examples as tutorials. Then we will revise the paper.
