### 2025-01-10 #7468 @SimonSekavcnik
Hi @danielskatz,

We have completed the benchmarks and various other comments. The documentation and the paper will be updated next week.
