### 2025-01-14 04be4ff
Updated the description of the lossy beamsplitter benchmark
