### 2025-01-14 52a371b
Added the State Management Benchmark
