### 2025-01-14 1d99e96
Added Jaynes-Cummings Model and Time Bin Encoding as a Tutorial openjournals/joss-reviews#7468
