+++ docs/examples/index.rst
@@ -1,0 +1,8 @@
+Examples
+========
+
+.. toctree::
+   :maxdepth: 2
+
+   time_bin_encoding
+   jaynes_cummings_model
