+++ docs/examples/index.rst
@@ -6,3 +6,4 @@ Examples
 
    time_bin_encoding
    jaynes_cummings_model
+   super_dense_coding
