### 2025-01-16 e7e7743
Pytest Passing, MyPy Passing
