+++ tests/state/test_composite_envelope.py
@@ -530,12 +530,8 @@ class TestCompositeEnvelopeMeasurementsVectors(unittest.TestCase):
         out = ce.measure(env1.fock, env2.fock, destructive=False)
         self.assertEqual(out[env1.fock], 1)
         self.assertEqual(out[env2.fock], 2)
-        self.assertEqual(out[env1.polarization], 0)
-        self.assertEqual(out[env2.polarization], 0)
         self.assertFalse(env1.fock.measured)
         self.assertFalse(env2.fock.measured)
-        self.assertFalse(env1.polarization.measured)
-        self.assertFalse(env2.polarization.measured)
         self.assertFalse(env1.measured)
         self.assertFalse(env2.measured)
         # Test when the envelopes are combined
@@ -544,14 +540,10 @@ class TestCompositeEnvelopeMeasurementsVectors(unittest.TestCase):
         out = ce.measure(env1.fock, env2.fock, destructive=False)
         self.assertEqual(out[env1.fock], 1)
         self.assertEqual(out[env2.fock], 2)
-        self.assertEqual(out[env1.polarization], 0)
-        self.assertEqual(out[env2.polarization], 0)
         self.assertFalse(env1.fock.measured)
         self.assertFalse(env2.fock.measured)
         self.assertFalse(env1.polarization.measured)
         self.assertFalse(env2.polarization.measured)
-        self.assertFalse(env1.measured)
-        self.assertFalse(env2.measured)
         self.assertEqual(env1.fock.state, 1)
         self.assertEqual(env2.fock.state, 2)
         self.assertEqual(env1.polarization.state, PolarizationLabel.H)
@@ -561,18 +553,12 @@ class TestCompositeEnvelopeMeasurementsVectors(unittest.TestCase):
         out = ce.measure(env1.fock, env2.fock, destructive=False)
         self.assertEqual(out[env1.fock], 1)
         self.assertEqual(out[env2.fock], 2)
-        self.assertEqual(out[env1.polarization], 0)
-        self.assertEqual(out[env2.polarization], 0)
         self.assertFalse(env1.fock.measured)
         self.assertFalse(env2.fock.measured)
-        self.assertFalse(env1.polarization.measured)
-        self.assertFalse(env2.polarization.measured)
         self.assertFalse(env1.measured)
         self.assertFalse(env2.measured)
         self.assertEqual(env1.fock.state, 1)
         self.assertEqual(env2.fock.state, 2)
-        self.assertEqual(env1.polarization.state, PolarizationLabel.H)
-        self.assertEqual(env2.polarization.state, PolarizationLabel.H)
 
         ce.combine(env1.fock, env2.fock)
         ce.combine(env1.polarization, env2.polarization)
@@ -580,18 +566,12 @@ class TestCompositeEnvelopeMeasurementsVectors(unittest.TestCase):
         out = ce.measure(env1.fock, env2.fock, destructive=False)
         self.assertEqual(out[env1.fock], 1)
         self.assertEqual(out[env2.fock], 2)
-        self.assertEqual(out[env1.polarization], 0)
-        self.assertEqual(out[env2.polarization], 0)
         self.assertFalse(env1.fock.measured)
         self.assertFalse(env2.fock.measured)
-        self.assertFalse(env1.polarization.measured)
-        self.assertFalse(env2.polarization.measured)
         self.assertFalse(env1.measured)
         self.assertFalse(env2.measured)
         self.assertEqual(env1.fock.state, 1)
         self.assertEqual(env2.fock.state, 2)
-        self.assertEqual(env1.polarization.state, PolarizationLabel.H)
[…]
