### 2025-01-16 #7468 @SimonSekavcnik
Dear @danielskatz,

I wanted to inform you that in my pull request, tqsd/photon_weave#9, I have incorporated all the updates to the code, benchmarks, and documentation. Please note that the documentation updates on the readthedocs will only be visible once the pull request is merged.

We are currently working on updating the paper and will soon provide you with a detailed list addressing each of the reviewers' comments individually.

We are planning to upload the updated paper tomorrow.
