### 2025-01-17 dc7466f
Fixed Polarization Contraction
