+++ examples/interoperable_devices/dense_decoder.py
@@ -8,19 +8,18 @@ from photon_weave.operation import Operation
 from photon_weave.operation.composite_operation import CompositeOperationType
 from photon_weave.operation.polarization_operation import PolarizationOperationType
 
+
 class DenseDecoder:
-
     def decode(self, env1: Envelope, env2: Envelope) -> tuple[int, int]:
-
         op_h = Operation(PolarizationOperationType.H)
         op_cnot = Operation(CompositeOperationType.CXPolarization)
 
         ce = CompositeEnvelope(env1, env2)
-        
+
         ce.apply_operation(op_cnot, env1.polarization, env2.polarization)
 
         env1.polarization.apply_operation(op_h)
-        
+
         m1 = env1.measure()
         m2 = env2.measure()
 
