+++ setup.py
@@ -1,9 +1,10 @@
 """fock envelope simulator"""
-from setuptools import find_packages, setup
 from pathlib import Path
 
+from setuptools import find_packages, setup
+
 this_directory = Path(__file__).parent
-long_description= (this_directory / "README.md").read_text()
+long_description = (this_directory / "README.md").read_text()
 
 setup(
     name="photon_weave",
@@ -12,12 +13,12 @@ setup(
     author_email="simon.sekavcnik@tum.de",
     description="General Quantum Simulator, with focus on optics",
     long_description=long_description,
-    long_description_content_type='text/markdown',
+    long_description_content_type="text/markdown",
     license="Apache 2.0",
     url="https://github.com/tqsd/photon_weave",
     project_urls={
         "Documentation": "https://photon-weave.readthedocs.io/en/latest/index.html",
-      },
+    },
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: Apache Software License",
