### 2025-01-17 27f425a
jitting some functions
