### 2025-01-17 ec4fe15
jitting kraus function
