### 2025-01-17 54e5311
jitting kron in combine
