+++ paper/paper.md
@@ -18,7 +18,7 @@ authors:
    orcid: 0000-0003-0091-3072
    affiliation: 1
 affiliations:
-  - name: Technical University of Munich, Theoretical Quantum System Design, Munich, Germany
+ - name: Technical University of Munich, Theoretical Quantum System Design, Munich, Germany
     index: 1
 date: 15.10.2024
 bibliography: paper.bib
@@ -26,28 +26,28 @@ bibliography: paper.bib
 # Summary
 Photon Weave is a quantum systems simulator designed to offer intuitive abstractions for simulating photonic quantum systems and their interactions in Fock spaces [@Fock1932] along with any custom Hilbert space. The simulator focuses on simplifying complex quantum state representations, such as continuous photonic states with polarization using envelopes that can mimic pulses, making it more approachable for specialized quantum simulations. While general-purpose quantum simulation libraries such as QuTiP [@johansson2012qutip] provide robust tools for quantum state manipulations, some require advanced software skills to manipulate complex system interactions. Photon Weave addresses this by abstracting such details, streamlining the simulation process, and allowing quantum systems to interact naturally as the simulation progresses.
 
-In contracts to other frameworks such as Qiskit [@aleksandrowicz2019qiskit], which are primarily designed for qubit-based computations, Photon Weave excels at simulating continuous-variable quantum systems, mainly photons, as well as custom quantum states that can interact dynamically. Furthermore, Photon Weave offers a balance of flexibility and automation by deferring the joining of quantum spaces using State Containers until necessary, enhancing computational efficiency. The simulator supports both CPU and GPU execution, ensuring scalability and performance for large-scale simulations. This is achieved by using the JAX [@jax2018github] library. As depicted in Fig. 1, you can see .... 
+In contracts to other frameworks such as Qiskit [@aleksandrowicz2019qiskit], which are primarily designed for qubit-based computations, Photon Weave excels at simulating continuous-variable quantum systems, mainly photons, as well as custom quantum states that can interact dynamically. Furthermore, Photon Weave offers a balance of flexibility and automation by deferring the joining of quantum spaces using State Containers until necessary, enhancing computational efficiency. The simulator supports CPU and GPU execution, ensuring scalability and performance for large-scale simulations. This is achieved by using the JAX [@jax2018github] library. As depicted in Fig. 1, you can see .... 
 
 # Statement of Need
-Tools like QuTiP, Qiskit, Piquasso, and Strawberry Fields [@kolarovszki2024piquassophotonicquantumcomputer; @killoran2019strawberry] already exist for modeling quantum phenomena, but many of them either require extensive user control (QuTiP) or enforce rigid circuit structures (Strawberry Fields). Researchers in quantum optics and related fields need a tool that simplifies photonic systems simulations and supports dynamic interactions between custom quantum systems. Photon Weave introduces such features without restricting itself to the circuit model so that researchers can focus on component development. Such a tool could be used to generate a library of components and gates that closely model real-world devices, fostering greater collaboration among scientists in those fields.
+Tools like QuTiP, Qiskit, Piquasso, and Strawberry Fields [@kolarovszki2024piquassophotonicquantumcomputer; @killoran2019strawberry] already exist for modeling quantum phenomena, but many of them either require extensive user control (QuTiP) or enforce rigid circuit structures (Strawberry Fields). Researchers in quantum optics and related fields need a tool that simplifies photonic systems simulations and supports dynamic interactions between custom quantum systems. Photon Weave introduces such features without restricting itself to the circuit model so researchers can focus on component development. Such a tool could generate a library of components and gates that closely model real-world devices, fostering greater collaboration among scientists in those fields.
 
 # Photon Weave Overview
[…]
