+++ paper/paper.md
@@ -29,7 +29,7 @@ Photon Weave is a quantum systems simulator designed to offer intuitive abstract
 In contracts to other frameworks such as Qiskit [@aleksandrowicz2019qiskit], which are primarily designed for qubit-based computations, Photon Weave excels at simulating continuous-variable quantum systems, mainly photons, as well as custom quantum states that can interact dynamically. Furthermore, Photon Weave offers a balance of flexibility and automation by deferring the joining of quantum spaces using State Containers until necessary, enhancing computational efficiency. The simulator supports CPU and GPU execution, ensuring scalability and performance for large-scale simulations. This is achieved by using the JAX [@jax2018github] library. As depicted in Fig. 1, you can see .... 
 
 # Statement of Need
-Tools like QuTiP, Qiskit, Piquasso, and Strawberry Fields [@kolarovszki2024piquassophotonicquantumcomputer; @killoran2019strawberry] already exist for modeling quantum phenomena, but many of them either require extensive user control (QuTiP) or enforce rigid circuit structures (Strawberry Fields). Researchers in quantum optics and related fields need a tool that simplifies photonic systems simulations and supports dynamic interactions between custom quantum systems. Photon Weave introduces such features without restricting itself to the circuit model so researchers can focus on component development. Such a tool could generate a library of components and gates that closely model real-world devices, fostering greater collaboration among scientists in those fields.
+Tools like QuTiP, Qiskit, Piquasso, and StrawberryFields [@kolarovszki2024piquassophotonicquantumcomputer; @killoran2019strawberry] already exist for modeling quantum phenomena, but many of them either require extensive user control (QuTiP) or enforce rigid circuit structures (StrawberryFields). Researchers in quantum optics and related fields need a tool that simplifies photonic systems simulations and supports dynamic interactions between custom quantum systems. Photon Weave introduces such features without restricting itself to the circuit model so researchers can focus on component development. Such a tool could generate a library of components and gates that closely model real-world devices, fostering greater collaboration among scientists in those fields.
 
 # Photon Weave Overview
 Photon Weave is a quantum simulation library designed for simulating any system, provided that simulating hardware meets the resource requirements. This simulator allows users to easily create, manipulate, and measure quantum systems.
@@ -44,22 +44,22 @@ Photon Weave's core functionality revolves around quantum state containers. Stat
 Photon Weave places a particular emphasis on the `Envelope` concept. An `Envelope` represents a pulse of light, where all photons are indistinguishable and share the same polarization, representing the $\mathcal{F}\otimes\mathcal{P}$ space where $\mathcal{F}$ represents the Fock space and $\mathcal{P}$ represents the Polarization space. Initially, when the states are separable, they are stored in the respective `Fock` and `Polarization` containers. In addition to the states, an `Envelope` holds essential metadata such as wavelength and temporal profile.
 
 ### Composite Envelopes
-When envelopes interact using a beam-splitter [@xiang2002theorem], their states must be joined. The necessary state data is extracted from their respective containers and tensorized into a product state in these cases. A `CompositeEnvelope` can contain multiple product spaces, which can be accessed from any of the contributing state containers. Additionally, `CompositeEnvelope` instances can be merged, allowing states within both envelopes to interact. Since any custom state can, in principle, interact with any other state, `CustomState` instances can also be included in a `CompositeEnvelope.`
[…]
