### 2025-01-20 c6729c0
Small update in the benchmark plotting
