+++ benchmarks/lossy_circuit/lossy_circuit.py
@@ -12,10 +12,10 @@ from cpuinfo import get_cpu_info
 import matplotlib.pyplot as plt
 
 SCRIPTS = {
-    "photon_weave": "lc_photon_weave.py",
-    "qutip": "lc_qutip.py",
-    "qiskit": "lc_qiskit.py",
-}
+    "photon_weave":"lc_photon_weave.py",
+    "qutip":"lc_qutip.py",
+    "qiskit": "lc_qiskit.py"
+    }
 
 system_info = {
     "OS": platform.system(),
@@ -50,24 +50,31 @@ def run_benchmark(initial_state):
                 stdout=subprocess.PIPE,
                 stderr=subprocess.PIPE,
                 text=True,
-                check=True,
-            )
+                check=True
+                )
             end_time = time.time()
             execution_time = end_time - start_time
 
             output = json.loads(process.stdout)
-            results[name] = {"time": execution_time, "output": output}
+            results[name] = {
+                "time": execution_time,
+                "output": output
+                }
         except subprocess.CalledProcessError as e:
             # Handle script execution errors
             print(f"Error running {name}: {e}")
-            results[name] = {"time": None, "output": None, "error": e.stderr}
+            results[name] = {
+                "time": None,
+                "output": None,
+                "error": e.stderr
+            }
         except json.JSONDecodeError as e:
             # Handle JSON parsing errors
             print(f"Error parsing JSON from {name}: {e}")
             results[name] = {
                 "time": None,
                 "output": None,
-                "error": "Invalid JSON output",
+                "error": "Invalid JSON output"
             }
     print(results)
     create_plot(results, None)
@@ -85,59 +92,81 @@ def add_system_info(fig, system_info_text):
     ax = fig.add_subplot(3, 1, 3)
     ax.axis("off")  # Turn off the axis
     ax.text(
-        0.5,
-        0.5,
-        system_info_text,
-        fontsize=12,
-        ha="center",
-        va="center",
-        bbox=dict(boxstyle="round", alpha=0.3),
+        0.5, 0.5, system_info_text, fontsize=12, ha="center", va="center",
+        bbox=dict(boxstyle="round", alpha=0.3)
     )
-
 
 def create_plot(data, file_location):
     linestyles = ("solid", "solid", "dotted")
     linecolors = ("black", "red", "blue")
     steps = list(range(1, 7))  # x-axis (steps)
     labels = list(data.keys())  # Method names
-    state_sizes = [data[label]["output"]["state_sizes"] for label in labels]
-    operator_sizes = [data[label]["output"]["operator_sizes"] for label in labels]
-    execution_times = [data[label]["time"] for label in labels]
+    state_sizes = [data[label]['output']['state_sizes'] for label in labels]
+    operator_sizes = [data[label]['output']['operator_sizes'] for label in labels]
+    execution_times = [data[label]['time'] for label in labels]
+
+    # Set up the figure with two side-by-side subplots
+    fig, (ax1, ax2) = plt.subplots(1, 2, figsize=(20, 5), constrained_layout=True)
+
+    # Left Plot: State Sizes (Logarithmic Scale)
+    for i, label in enumerate(labels):
+        ax1.plot(steps, state_sizes[i], marker='o',
+                 label=f"{label} ({execution_times[i]:.2f}s)",
[…]
