+++ paper/paper.md
@@ -24,7 +24,7 @@ date: 15.10.2024
 bibliography: paper.bib
 ---
 # Summary
-Photon Weave is a quantum systems simulator designed to offer intuitive abstractions for simulating photonic quantum systems and their interactions in Fock spaces [@Fock1932] along with any custom Hilbert space. The simulator focuses on simplifying complex quantum state representations, such as continuous photonic states with polarization using envelopes that can mimic pulses, making it more approachable for specialized quantum simulations. While general-purpose quantum simulation libraries such as QuTiP [@johansson2012qutip] provide robust tools for quantum state manipulations, some require advanced software skills to manipulate complex system interactions. Photon Weave addresses this by abstracting such details, streamlining the simulation process, and allowing quantum systems to interact naturally as the simulation progresses.
+Photon Weave is a quantum systems simulator designed to offer intuitive abstractions for simulating quantum photonic systems and their interactions in Fock spaces [@Fock1932] along with any custom Hilbert spaces. The simulator focuses on simplifying complex quantum state representations, such as continuous photonic states with polarization using envelopes that can mimic pulses, making it more approachable for specific quantum simulations. While general-purpose quantum simulation libraries such as QuTiP [@johansson2012qutip] provide robust tools for quantum state manipulations, some require advanced software skills to manipulate complex system interactions. Photon Weave addresses this by abstracting such details, streamlining the simulation process, and allowing quantum systems to interact naturally as the simulation progresses.
 
 In contracts to other frameworks such as Qiskit [@aleksandrowicz2019qiskit], which are primarily designed for qubit-based computations, Photon Weave excels at simulating continuous-variable quantum systems, mainly photons, as well as custom quantum states that can interact dynamically. Furthermore, Photon Weave offers a balance of flexibility and automation by deferring the joining of quantum spaces using State Containers until necessary, enhancing computational efficiency. The simulator supports CPU and GPU execution, ensuring scalability and performance for large-scale simulations. This is achieved by using the JAX [@jax2018github] library. As depicted in Fig. 1, you can see .... 
 
