+++ paper/paper.md
@@ -29,7 +29,9 @@ Photon Weave is a quantum systems simulator designed to offer intuitive abstract
 In contracts to other frameworks such as Qiskit [@aleksandrowicz2019qiskit], which are primarily designed for qubit-based computations, Photon Weave excels at simulating continuous-variable quantum systems, mainly photons, as well as custom quantum states that can interact dynamically. Furthermore, Photon Weave offers a balance of flexibility and automation by deferring the joining of quantum spaces using State Containers until necessary, enhancing computational efficiency. The simulator supports CPU and GPU execution, ensuring scalability and performance for large-scale simulations. This is achieved by using the JAX [@jax2018github] library.
 
 # Statement of Need
-Tools like QuTiP, Qiskit, Piquasso, and StrawberryFields [@kolarovszki2024piquassophotonicquantumcomputer; @killoran2019strawberry] already exist for modeling quantum phenomena, but many of them either require extensive user control (QuTiP) or enforce rigid circuit structures (StrawberryFields). Researchers in quantum optics and related fields need a tool that simplifies photonic systems simulations and supports dynamic interactions between custom quantum systems. Photon Weave introduces such features without restricting itself to the circuit model so researchers can focus on component development. Such a tool could generate a library of components and gates that closely model real-world devices, fostering greater collaboration among scientists in those fields. Fig.1 depicts a complicated scenario where every Beam Splitter is considered lossy and the system evolution is tracked with each step. ![Figure 1. Simulation of lossy Mach Zender Interferometer. The simulation tracks the state evolution throughout the experiment. The losses here are photon absorption.](benchmarks/lossy_circuit/circuit.png){label="fig.1"} In Fig.2, You can see the performance regarding speed where the steps represent the number of applied operations. ![Figure 2. Comparison between Photon Weave, Qiskit, and QuTip regarding simulation time and the required space to simulate the experiment in \ref{fig.1}](lossy_circuit_paper-2.png)
+Tools like QuTiP, Qiskit, Piquasso, and StrawberryFields [@kolarovszki2024piquassophotonicquantumcomputer; @killoran2019strawberry] already exist for modeling quantum phenomena, but many of them either require extensive user control (QuTiP) or enforce rigid circuit structures (StrawberryFields). Researchers in quantum optics and related fields need a tool that simplifies photonic systems simulations and supports dynamic interactions between custom quantum systems. Photon Weave introduces such features without restricting itself to the circuit model so researchers can focus on component development. Such a tool could generate a library of components and gates that closely model real-world devices, fostering greater collaboration among scientists in those fields.
+
+Fig.1 depicts a complicated scenario where every Beam Splitter is considered lossy and the system evolution is tracked with each step. ![Figure 1. Simulation of lossy Mach Zender Interferometer. The simulation tracks the state evolution throughout the experiment. The losses here are photon absorption.](circuit.png){label="fig.1"}. In Fig.2, You can see the performance regarding speed where the steps represent the number of applied operations. ![Figure 2. Comparison between Photon Weave, Qiskit, and QuTip regarding simulation time and the required space to simulate the experiment in \ref{fig.1}](lossy_circuit_paper-2.png)
 
 # Photon Weave Overview
 Photon Weave is a quantum simulation library designed for simulating any system, provided that simulating hardware meets the resource requirements. This simulator allows users to easily create, manipulate, and measure quantum systems.
