+++ tox.ini
@@ -21,8 +21,8 @@ deps =
     isort
 commands =
     flake8 photon_weave
-    black photon_weave
     isort photon_weave
+    black -l 88 photon_weave
 
 
 

[…]
