### 2025-01-22 7bdc178
update date of lic and conf.py in docs
