### 2025-01-23 #7468 @taladjidi
Hi, I'll have another look at this, but I won't be able to do it until monday, so I'll try to post my eventual comments monday night (Paris time). Thanks for the work @SimonSekavcnik !
