### 2025-01-27 #7468 @a-eghrari
Hello, I looked at the paper, and it seems acceptable. However, I would need more time to review the examples in the documentation. My only comment for now:

* I can see community and contribution guidelines in the repository. I would also suggest adding them to the documentation.
