### 2025-01-30 #7468 @danielskatz
👋 @taladjidi - You have an existing checklist at https://github.com/openjournals/joss-reviews/issues/7468#issuecomment-2493239250
And I see you have generated a new one.
Can you say why?
