### 2025-02-11 #7468 @danielskatz
@taladjidi - Can you now complete [your checklist](https://github.com/openjournals/joss-reviews/issues/7468#issuecomment-2493239250)?  If not, can you say what else you think needs to be done?
