### 2025-02-11 #7468 @danielskatz
👋 @a-eghrari - Can you say what you think still needs to be done for you to check off the items on [your checklist](https://github.com/openjournals/joss-reviews/issues/7468#issuecomment-2500216442)?
