### 2025-02-17 #7468 @danielskatz
👋 @taladjidi - I just three more items in [your checklist](https://github.com/openjournals/joss-reviews/issues/7468#issuecomment-2493239250). What is needed for you to be able to check them off?
