### 2025-02-17 #7468 @danielskatz
👋 @SimonSekavcnik - Can you address @a-eghrari's point above, about the lack of community guidelines in the documentation? (I note that this is also one of the unchecked items for @taladjidi)
