### 2025-02-18 bfe8e9f
updating docs to include code of conduct and contributions
