+++ docs/requirements.txt
@@ -1,3 +1,4 @@
 sphinx 
 sphinx-rtd-theme 
 sphinx-autodoc-typehints
+myst-parser
