### 2025-02-18 e759c3f
updating docs to include code of conduct and contributions (#12)
updating docs with code of conduct and contribution
