+++ docs/contributing.rst
@@ -1,0 +1,49 @@
+.. _contributing:
+
+Contributing to Photon Weave
+====================
+
+We love your input! We want to make contributing to this project as easy and transparent as possible, whether it's:
+- Reporting a bug
+- Discussing the current state of the code
+- Submitting a fix
+- Proposing new features
+- Becoming a maintainer
+
+We Develop with Github
+====================
+We use github to host code, to track issues and feature requests, as well as accept pull requests.
+
+we use [Github Flow](https://guides.github.com/introduction/flow/index.html), All Code Changes Happen Through Pull Requests
+Pull requests are the best way to propose changes to the codebase (we use [Github Flow](https://guides.github.com/introduction/flow/index.html). We actively welcome your pull requests:
+1. Fork the repo and create your branch from `master`.
+2. If you've added code that should be tested, add tests.
+3. If you've changed APIs, update the documentation.
+4. Ensure the test suite passes.
+5. Make sure your code lints.
+6. Issue the pull request.
+
+Any Contribution you make will be under the Apache-2.0 License
+
+In short, when you submit code changes, your submissions are understood to bu under the same [Apache-2.0 License](https://www.apache.org/licenses/LICENSE-2.0) that covers the project. Feel free to contact the maintainers if that's a concern.
+
+Report bugs using Github's [issues](https://github.com/tqsd/photon_weave/issues)
+====================
+We use GitHub issues to track public bugs. Report a bug by opening a new issue.
+
+Write bug reports with detail, background and sample code
+**Great Bug Reports** tend to have:
+- A quick summary and/or background
+- Steps to reproduce
+  - Be specific!
+  - Give sample code if you can.
+- What you expected would happen.
+- What actually happens.
+- Notes (possibly including why you think this might be happening, or stuff you tried that didn't work)
+
+Use a Consistent Coding Stype
+We aim to be consistent with PEP8.
+
+License
+=========
+By contributing, you agree that your contributions will be licensed under Apache-2.0 License.
