+++ docs/code_of_conduct.rst
@@ -1,11 +1,9 @@
-.. _code_of_conduct:
-
 Code of Conduct
 ===============
 
 Contributor Covenant Code of Conduct
-================
-Our Pledge
+
+**Our Pledge**
 
 We as members, contributors, and leaders pledge to make participation in our
 community a harassment-free experience for everyone, regardless of age, body
@@ -17,8 +15,8 @@ and orientation.
 We pledge to act and interact in ways that contribute to an open, welcoming,
 diverse, inclusive, and healthy community.
 
-Our Standards
-====================
+**Our Standards**
+
 Examples of behavior that contributes to a positive environment for our
 community include:
 
@@ -41,8 +39,8 @@ Examples of unacceptable behavior include:
 * Other conduct which could reasonably be considered inappropriate in a
   professional setting
 
-Enforcement Responsibilities
-====================
+**Enforcement Responsibilities**
+
 
 Community leaders are responsible for clarifying and enforcing our standards of
 acceptable behavior and will take appropriate and fair corrective action in
@@ -54,16 +52,16 @@ comments, commits, code, wiki edits, issues, and other contributions that are
 not aligned to this Code of Conduct, and will communicate reasons for moderation
 decisions when appropriate.
 
-Scope
-=======
+**Scope**
+
 This Code of Conduct applies within all community spaces, and also applies when
 an individual is officially representing the community in public spaces.
 Examples of representing our community include using an official e-mail address,
 posting via an official social media account, or acting as an appointed
 representative at an online or offline event.
 
-Enforcement
-====================
+**Enforcement**
+
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
 reported to the community leaders responsible for enforcement at
 simon.sekavcnik@tum.de.
@@ -72,8 +70,8 @@ All complaints will be reviewed and investigated promptly and fairly.
 All community leaders are obligated to respect the privacy and security of the
 reporter of any incident.
 
-Enforcement Guidelines
-====================
+**Enforcement Guidelines**
+
 Community leaders will follow these Community Impact Guidelines in determining
 the consequences for any action they deem in violation of this Code of Conduct:
 
@@ -118,8 +116,8 @@ individual, or aggression toward or disparagement of classes of individuals.
 **Consequence**: A permanent ban from any sort of public interaction within
 the community.
 
-Attribution
-===========
+**Attribution**
+
 This Code of Conduct is adapted from the [Contributor Covenant][homepage],
 version 2.0, available at
 https://www.contributor-covenant.org/version/2/0/code_of_conduct.html.
