### 2025-02-18 #7468 @kareem1925
> Hello, I looked at the paper, and it seems acceptable. However, I would need more time to review the examples in the documentation. My only comment for now:
>
> * I can see community and contribution guidelines in the repository. I would also suggest adding them to the documentation.

**_Reply_**: I believe we have updated documentation as you can see [here](https://photon-weave.readthedocs.io/en/latest/code_of_conduct.html#)
