+++ paper/paper.md
@@ -31,7 +31,7 @@ PhotonWeave is a quantum systems simulator designed to offer intuitive abstracti
 In contracts to other frameworks such as Qiskit [@aleksandrowicz2019qiskit], which are primarily designed for qubit-based computations, PhotonWeave excels at simulating continuous-variable quantum systems, mainly photons, as well as custom quantum states that can interact dynamically. Furthermore, PhotonWeave offers a balance of flexibility and automation by deferring the joining of quantum spaces using State Containers until necessary, enhancing computational efficiency. The simulator supports CPU and GPU execution, ensuring scalability and performance for large-scale simulations. This is achieved by using the JAX [@jax2018github] library.
 
 # Statement of Need
-Tools like QuTiP, Qiskit, Piquasso, and StrawberryFields [@kolarovszki2024piquassophotonicquantumcomputer; @killoran2019strawberry] already exist for modeling quantum phenomena, but many of them either require extensive user control (QuTiP) or enforce rigid circuit structures (StrawberryFields). Researchers in quantum optics and related fields need a tool that simplifies photonic systems simulations and supports dynamic interactions between custom quantum systems. PhotonWeave introduces such features without restricting itself to the circuit model so researchers can focus on component development. Such a tool could generate a library of components and gates that closely model real-world devices, fostering greater collaboration among scientists in those fields. In Fig. \ref{fig:1}, a complicated scenario of lossy Beam Splitters is depicted, and in Fig. \ref{fig:2}, the performance is compared to Qiskit and QuTip.
+Tools like QuTiP, Qiskit, Piquasso, StrawberryFields [@kolarovszki2024piquassophotonicquantumcomputer; @killoran2019strawberry], and SOQCS [@osca2023soqcs] already exist for modeling quantum phenomena. However, many of them either require extensive user control (QuTiP) or enforce rigid circuit structures (StrawberryFields, SOQCS). Researchers in quantum optics and related fields need a tool that simplifies photonic systems simulations and supports dynamic interactions between custom quantum systems. PhotonWeave introduces such features without restricting itself to the circuit model, allowing researchers to focus on component development. Such a tool could generate a library of components and gates that closely model real-world devices, fostering greater collaboration among scientists in these fields. In Fig. \ref{fig:1}, a complex scenario of lossy beam splitters is depicted, and in Fig. \ref{fig:2}, the performance is compared to Qiskit and QuTiP.
 
 
 ![The simulation of lossy Beam Splitters. The simulation tracks the state evolution throughout the experiment. The losses here are photon absorption. \label{fig:1}](circuit.png)
