+++ paper/paper.bib
@@ -82,3 +82,15 @@
   year={2002},
   publisher={APS}
 }
+
+@article{Osca_2024,
+   title={SOQCS: A Stochastic Optical Quantum Circuit Simulator},
+   volume={25},
+   ISSN={2352-7110},
+   url={http://dx.doi.org/10.1016/j.softx.2023.101603},
+   DOI={10.1016/j.softx.2023.101603},
+   journal={SoftwareX},
+   publisher={Elsevier BV},
+   author={Osca, Javier and Vala, Jiri},
+   year={2024},
+   month=feb, pages={101603} }
