+++ paper/paper.bib
@@ -26,14 +26,7 @@
   year={2021},
   publisher={IEEE}
 }
-@inproceedings{wille2019ibm,
-  title={IBM’s Qiskit tool chain: Working with and developing for real quantum computers},
-  author={Wille, Robert and Van Meter, Rod and Naveh, Yehuda},
-  booktitle={2019 Design, Automation \& Test in Europe Conference \& Exhibition (DATE)},
-  pages={1234--1240},
-  year={2019},
-  organization={IEEE}
-}
+
 @software{jax2018github,
   author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
