### 2025-02-18 #7468 @kareem1925
> > Dear [@danielskatz](https://github.com/danielskatz), [@a-eghrari](https://github.com/a-eghrari) and [@taladjidi](https://github.com/taladjidi),
> > in the merge above, we have included all of the updates, including the updates to the paper. The [documentation](https://photon-weave.readthedocs.io/en/latest/index.html) was also updated. We want to sincerely thank both reviewers, as we believe that the quality of our package has already increased due to your suggestions and remarks.
> > ## Addressing Comments by Taladjidi
>
> > ```
> > * How does your package compare to SOQCS ?
> >
> >   * We created the comparison (attached below) but did not include it in the paper or the repository.
> >   * [comparison (1).pdf](https://github.com/user-attachments/files/18491172/comparison.1.pdf)
> > ```
>
> I think it should be in the paper, along with comparisons to other framework, at least in one sentence in the statement of need. Furthermore, while your benchmark shows that your code is sparser in operator sizes, in the end it consumes more memory. If this is due to Jax caching, this can be estimated in order to show that the inherent memory use of photon weave is indeed better. I guess it should be possible to only benchmark the execution after some dry runs. I don't know for JAX, but I know that for Numba it is possible ([this looks like a good start](https://jax.readthedocs.io/en/latest/profiling.html)). Your code is much faster that qiskit qutip, but it is hidden in the legend of the plot, you could show it more. I think and improved version of this paper could fit well in the paper as to visually highlight the appeal of your package.
>
> Some citations for Qutip / Qiskit / Strawberry fields would be welcome.
>
> The contribution guidelines are still missing but otherwise I think most of my concerns have been addressed !
>
> Thanks for the good work [@SimonSekavcnik](https://github.com/SimonSekavcnik)

**_Reply_**: We have updated the statement of need and included SOQCS, and we kept it at a minimum to adhere to the number of permitted words. <ins> Regarding the memory simulation </ins>: The memory profiler in JAX is used for GPU memory profiling, and it doesn't profile the CPU executions. So, we decided to stick to the same approach we had chosen, and we tried ignoring the first 5 iterations as they serve as a warm or for compiling the operations, and the new results followed the same trend with no changes. But we are open to any new suggestions from your side :)
