+++ docs/contributing.rst
@@ -12,8 +12,8 @@ We love your input! We want to make contributing to this project as easy and tra
 
 We use github to host code, to track issues and feature requests, as well as accept pull requests.
 
-we use [Github Flow](https://guides.github.com/introduction/flow/index.html), All Code Changes Happen Through Pull Requests
-Pull requests are the best way to propose changes to the codebase (we use [Github Flow](https://guides.github.com/introduction/flow/index.html). We actively welcome your pull requests:
+we use `Github Flow <https://guides.github.com/introduction/flow/index.html>`_, All Code Changes Happen Through Pull Requests
+Pull requests are the best way to propose changes to the codebase (we use Github Flow). We actively welcome your pull requests:
 1. Fork the repo and create your branch from `master`.
 2. If you've added code that should be tested, add tests.
 3. If you've changed APIs, update the documentation.
@@ -23,9 +23,9 @@ Pull requests are the best way to propose changes to the codebase (we use [Githu
 
 Any Contribution you make will be under the Apache-2.0 License
 
-In short, when you submit code changes, your submissions are understood to bu under the same [Apache-2.0 License](https://www.apache.org/licenses/LICENSE-2.0) that covers the project. Feel free to contact the maintainers if that's a concern.
+In short, when you submit code changes, your submissions are understood to bu under the same `Apache-2.0 Licens <https://www.apache.org/licenses/LICENSE-2.0>`_ that covers the project. Feel free to contact the maintainers if that's a concern.
 
-Report bugs using Github's [issues](https://github.com/tqsd/photon_weave/issues)
+Report bugs using Github's `issues <https://github.com/tqsd/photon_weave/issues>`_
 
 We use GitHub issues to track public bugs. Report a bug by opening a new issue.
 
