+++ docs/contributing.rst
@@ -14,11 +14,17 @@ We use github to host code, to track issues and feature requests, as well as acc
 
 we use `Github Flow <https://guides.github.com/introduction/flow/index.html>`_, All Code Changes Happen Through Pull Requests
 Pull requests are the best way to propose changes to the codebase (we use Github Flow). We actively welcome your pull requests:
+
 1. Fork the repo and create your branch from `master`.
+
 2. If you've added code that should be tested, add tests.
+
 3. If you've changed APIs, update the documentation.
+
 4. Ensure the test suite passes.
+
 5. Make sure your code lints.
+
 6. Issue the pull request.
 
 Any Contribution you make will be under the Apache-2.0 License
