### 2025-02-18 #7468 @SimonSekavcnik
In addition to what @kareem1925 has said with regards to the profiling, the problem is that Jax library is quite memory intensive, and we couldn't come up with a way of isolating the state and operator sizes from the whole execution. We could calculate the sizes as we did in the lossy beam splitter benchmark, but even than this benchmark is designed to not be favorable to our tool in terms of memory consumption. Still it shows that our tool is faster than the two tools compared to.
