### 2025-02-18 #7468 @danielskatz
> Dear [@danielskatz](https://github.com/danielskatz) , I have additional question about the paper. We would like to include longer acknowledgement would this be possible even if it goes beyond the word count limit?

Yes, as long as it's not too long - basically a paragraph is fine.
