### 2025-02-18 196c190
Updated acknowledgments paper.md
