### 2025-02-19 077d9af
Updated ACK in paper.md
