### 2025-02-24 #7468 @SimonSekavcnik
Dear @danielskatz ,
I have updated the acknowledgements. Let me know if anything else is needed from my side.
I think that both lists are totally checked off.
