### 2025-02-25 #7468 @danielskatz
In the acknowledgments, there's an extra period at the end of the first sentence. Also, the second sentence is not a complete sentence. Should it be combined with the first sentence? If not, can you make it a proper sentence.
