+++ paper/paper.bib
@@ -1,30 +1,39 @@
 @article{johansson2012qutip,
   title={QuTiP: An open-source Python framework for the dynamics of open quantum systems},
-  author={Johansson, J Robert and Nation, Paul D and Nori, Franco},
-  journal={Computer physics communications},
-  volume={183},
-  number={8},
-  pages={1760--1772},
-  year={2012},
-  publisher={Elsevier}
-}
+   volume={183},
+   ISSN={0010-4655},
+   url={http://dx.doi.org/10.1016/j.cpc.2012.02.021},
+   DOI={10.1016/j.cpc.2012.02.021},
+   number={8},
+   journal={Computer Physics Communications},
+   publisher={Elsevier BV},
+   author={Johansson, J.R. and Nation, P.D. and Nori, Franco},
+   year={2012},
+   month=aug, pages={1760–1772} }
+
 @article{killoran2019strawberry,
-  title={Strawberry fields: A software platform for photonic quantum computing},
-  author={Killoran, Nathan and Izaac, Josh and Quesada, Nicol{\'a}s and Bergholm, Ville and Amy, Matthew and Weedbrook, Christian},
-  journal={Quantum},
-  volume={3},
-  pages={129},
-  year={2019},
-  publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften}
+  title={Strawberry Fields: A Software Platform for Photonic Quantum Computing},
+   volume={3},
+   ISSN={2521-327X},
+   url={http://dx.doi.org/10.22331/q-2019-03-11-129},
+   DOI={10.22331/q-2019-03-11-129},
+   journal={Quantum},
+   publisher={Verein zur Forderung des Open Access Publizierens in den Quantenwissenschaften},
+   author={Killoran, Nathan and Izaac, Josh and Quesada, Nicolás and Bergholm, Ville and Amy, Matthew and Weedbrook, Christian},
+   year={2019},
+   month=mar, pages={129} 
 }
 @article{diadamo2021qunetsim,
-  title={Qunetsim: A software framework for quantum networks},
-  author={DiAdamo, Stephen and N{\"o}tzel, Janis and Zanger, Benjamin and Be{\c{s}}e, Mehmet Mert},
-  journal={IEEE Transactions on Quantum Engineering},
+  title={QuNetSim: A Software Framework for Quantum Networks},
   volume={2},
-  pages={1--12},
+  ISSN={2689-1808},
+  url={http://dx.doi.org/10.1109/TQE.2021.3092395},
+  DOI={10.1109/tqe.2021.3092395},
+  journal={IEEE Transactions on Quantum Engineering},
+  publisher={Institute of Electrical and Electronics Engineers (IEEE)},
+  author={Diadamo, Stephen and Nötzel, Janis and Zanger, Benjamin and Beşe, Mehmet Mert},
   year={2021},
-  publisher={IEEE}
+  pages={1–12}
 }
 
 @software{jax2018github,
@@ -50,10 +59,15 @@
   pages = {622–647}
 }
 
-@article{aleksandrowicz2019qiskit,
-  title={Qiskit: An open-source framework for quantum computing},
-  author={Aleksandrowicz, Gadi and others},
-  year={2019}
+@misc{javadiabhari2024quantumcomputingqiskit,
+      title={Quantum computing with Qiskit}, 
+      author={Ali Javadi-Abhari and Matthew Treinish and Kevin Krsulich and Christopher J. Wood and Jake Lishman and Julien Gacon and Simon Martiel and Paul D. Nation and Lev S. Bishop and Andrew W. Cross and Blake R. Johnson and Jay M. Gambetta},
[…]
