### 2025-02-26 #7468 @SimonSekavcnik
Dear @danielskatz ,
unfortunately we cannot change this part of the acknowledgments as we are required to write in a specific way.
Please see:  [6G-Life Ack](https://6g-life.de/ack/)
