### 2025-02-27 #7468 @kareem1925
@danielskatz
> The tagged release is: https://github.com/tqsd/photon_weave/releases/tag/0.1.6
> The archived version on Zenodo is: https://doi.org/10.5281/zenodo.14932826
