### 2025-03-01 #7468 @danielskatz
👋 @SimonSekavcnik & @kareem1925 - I'll repeat one of the instructions above:

>Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

Please fix the title.
