### 2025-03-01 #7468 @danielskatz
👋 @SimonSekavcnik & @kareem1925 - In addition to the zenodo request above, please check the proof that this command will generate and confirm that it looks ok.
