### 2025-03-01 #7468 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
