### 2025-03-03 #7468 @danielskatz
@kareem1925 - thanks.  Editorialbot commands need to be the first thing in a new comment, but also, you don't have permission to run this command, it's something I need to do.
