### 2025-03-03 #7468 @danielskatz
Also, it seems that the Zenodo metadata (title) is not changed yet. Please let me know when that is done.
