### 2025-03-03 #7468 @danielskatz
Congratulations to @SimonSekavcnik (Simon Sekavcnik), @kareem1925, and Janis Nötzel on your publication!!

And thanks to @taladjidi & @a-eghrari for reviewing!
JOSS depends on volunteers and couldn't do this without you
