### 2025-03-03 #7468 @SimonSekavcnik
I'm thrilled to see that our package was published in the JOSS! 🎉

I would like to sincerely thank the editor, @danielskatz, and the reviewers, @taladjidi and @a-eghrari, for their valuable feedback and insights, which have helped improve the quality of our work.

A special thank you to my co-authors, @kareem1925 and @janisCGN, for their collaboration and contributions throughout this process. 🙌

I'm excited to see what the future holds for PhotonWeave! 🚀✨
